resistance (simplified)
The equivalent resistance of resistors connected in series is calculated by adding
The resistances of the individual resistors. Req = R1 + R2
The formula for resistors connected in parallel is a bit more complicated. Given two
Resistors with resistors R1 and R2 are connected in parallel with an equivalent resistance
Given by the inverse of the reciprocal sum:
(1 / R1) + (1 / R2)
Write a program that gives two resistors (R1, R2) the product of the equivalent resistance (Req)
When they are connected in series and when they are connected in parallel.